requirements.txt
requests
spotipy
beautifulsoup4
python-dotenv